move inspector page picking code out of metro.ts
